skin-cancer-detection/requirements.txt
torch>=1.7.0
torchvision>=0.8.1
opencv-python>=4.4.0
numpy>=1.19.4
matplotlib>=3.3.3
albumentations>=0.5.2
PyYAML>=5.3.1
scikit-learn>=0.24.0
pandas>=1.1.5